#1400. File delete/rename commit
