+++ LanguageFeatures/Inline-classes/static_analysis_inline_class_A09_t03.dart
