+++ LanguageFeatures/Inline-classes/static_analysis_inline_class_A10_t03.dart
